Patched Tuple Error
Patched the Tuple error, made it so when the batch files crash, it will instantly restart when you press enter.

I also updated README.md to be the Official Fantasy version! I also removed PPYTHON_PATH from .gitignore so the game will run when you pull the source to run the game!