Updated Options Menu
I know it's being replaced. So, I thought it might be some good practice to resize the stuff like the Options page to so it's nearly evenly placed. I also replaced any text regarding Goofy Speedway in the Toon Tips.

I also resized some of the buttons and fixed a text bug that made the "Discord Rich Presence: Off" appear as so: "Discord Rich Presence : Off".