I added DiscordRPC
I added DiscordRPC from other branches but since this branch is still public, I moved the ClientID to a json file that you have to make yourself. So, I've disabled Rich Presence by default. Though, it should be incredibly simple to enable and set up yourself. But it should work.