Added standalone launcher script
I added a standalone launcher script to the "scripts" folder which will just launch the client and nothing else (this is in case you need to exclusively close the client batch). I also updated the notice on the game's launcher to more accurately reflect how the game's account system works. There are also more quality of life features with the root game launcher!