Overhauled the start_game.bat
Overhauled the start_game.bat to make it more accessible to boot the game by checking if the requirements to launch the game are all installed. If they are not, it will install all of the requirements. It also checks for updates on the repository and the newly fixed-up submodules feature then automatically installs any updates it may have.