Return of the Toon News For The Amused!
This update sees the return of the "Toon News For The Amused!" page! Most of the functionality has been returned with 95 of the 420 images being missing from the list (spanning 2010, 2012, and 2013). This update also updates the DiscordRPC's clientID, patches a crash in the IssueFrame.py file, and restores the functionality of the "scripts/start_standalone_game.bat" file!